Moved functions that draw lines from point a to b into seperate file
