+++ vpr/src/draw/draw.h
@@ -55,19 +55,11 @@ ezgl::point2d atom_pin_draw_coord(AtomPinId pin);
 //Returns the drawing coordinates of the specified tnode
 ezgl::point2d tnode_draw_coord(tatum::NodeId node);
 
-void draw_mux_with_size(ezgl::point2d origin, e_side orientation, float height, int size, ezgl::renderer* g);
-void draw_chany_to_chany_edge(RRNodeId from_node, RRNodeId to_node, int to_track, short switch_type, ezgl::renderer* g);
-void draw_chanx_to_chanx_edge(RRNodeId from_node, RRNodeId to_node, int to_track, short switch_type, ezgl::renderer* g);
-void draw_chanx_to_chany_edge(int chanx_node, int chanx_track, int chany_node, int chany_track, enum e_edge_dir edge_dir, short switch_type, ezgl::renderer* g);
-void draw_pin_to_pin(int opin, int ipin, ezgl::renderer* g);
-void draw_pin_to_sink(int ipin_node, int sink_node, ezgl::renderer* g);
-void draw_source_to_pin(int source_node, int opin_node, ezgl::renderer* g);
-void draw_pin_to_chan_edge(int pin_node, int chan_node, ezgl::renderer* g);
-
 void annotate_draw_rr_node_costs(ClusterNetId net, int sink_rr_node);
 void clear_draw_rr_annotations();
 
 ezgl::color to_ezgl_color(vtr::Color<float> color);
+void draw_mux_with_size(ezgl::point2d origin, e_side orientation, float height, int size, ezgl::renderer* g);
 
 void draw_screen();
